+++ DataGateway.Service/Azure.DataGateway.Service.csproj
@@ -73,9 +73,4 @@
   <ItemGroup>
     <ProjectCapability Include="CSharp;Managed;ClassDesigner" />
   </ItemGroup>
-  <ItemGroup>
-    <Content Update="appsettings.Cosmos.json">
-      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
-    </Content>
-  </ItemGroup>
 </Project>
