+++ DataGateway.Service/appsettings.json
@@ -6,5 +6,16 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "DataGatewayConfig": {
+    "DatabaseType": "",
+    "DatabaseConnection": {
+      "ServerEndpointUrl": "",
+      "AuthorizationKey": "",
+      "Server": "",
+      "Database": "",
+      "Container": "",
+      "ConnectionString": ""
+    }
+  }
 }
